Refactor: UserChatId - primary key changed to new field id NotificationSchedulerImpl - renamed + code refactor RentalRepository - @Param removed UserChatIdRepository - method findByUserId renamed to findUserByTelegramId
