+++ src/main/java/online/carsharing/entity/UserChatId.java
@@ -2,6 +2,8 @@ package online.carsharing.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
@@ -13,8 +15,11 @@ import lombok.Setter;
 @Setter
 public class UserChatId {
     @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
     @Column(name = "user_telegram_id", nullable = false)
-    private Long userId;
+    private Long userTelegramId;
 
     @Column(name = "chat_id", nullable = false)
     private Long chatId;
