+++ src/main/java/online/carsharing/repository/rental/RentalRepository.java
@@ -5,7 +5,6 @@ import java.util.List;
 import online.carsharing.entity.Rental;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 public interface RentalRepository extends JpaRepository<Rental, Long> {
     @Query("SELECT r FROM Rental r JOIN FETCH r.car "
@@ -17,5 +16,5 @@ public interface RentalRepository extends JpaRepository<Rental, Long> {
 
     @Query("SELECT r FROM Rental r LEFT JOIN FETCH r.car LEFT JOIN FETCH"
             + " r.user WHERE r.returnDate <= :tomorrow AND r.actualReturnDate IS NULL")
-    List<Rental> findOverdueRentals(@Param("tomorrow") LocalDate tomorrow);
+    List<Rental> findOverdueRentals(LocalDate tomorrow);
 }
