+++ src/main/resources/db/changelog/changes/vyshynskyi/create-users-chat_id-table.yaml
@@ -7,10 +7,16 @@ databaseChangeLog:
             tableName: users_chat_id
             columns:
               - column:
+                  name: id
+                  type: bigint
+                  autoIncrement: true
+                  constraints:
+                    primaryKey: true
+              - column:
                   name: user_telegram_id
                   type: bigint
                   constraints:
-                    primaryKey: true
+                    nullable: false
               - column:
                   name: chat_id
                   type: bigint
